// src/WebUI/ClientApp_Old/src/app/enums/pathname.ts
export enum Pathname {
  Home = "Home",
  Buildings = "Buildings",
  Rooms = "Rooms",
  RoomExtraFields = "RoomExtraFields",
  NewBooking = "NewBooking",
  MyBookings = "MyBookings",
  GlobalList = "GlobalList",
  GlobalListName = "GlobalListName",
  AdminMenu = "AdminMenu",
  RoomGroup = "RoomGroup",
  MyInfo = "MyInfo",
  DraftPaper = "DraftPaper",
  PermissionGroup = "PermissionGroup",
  UserGroup = "UserGroup",
  UserRole = "UserRole",
  AllSimpleUserInfoList = "AllSimpleUserInfoList",
  UserOrgChart = "UserOrgChart",
  ClosedDay = "ClosedDay",
  ClosedDayList = "ClosedDayList",
  SmtpSetting = "SmtpSetting",
  SimpleUserInfoList = "SimpleUserInfoList",
  RoomReservation = "RoomReservation",
}